refactor(producers): lift recall-input helpers out of the registry file
The recall-attribution wiring grew producers/index.ts from 347 to 535 lines,
past the 500-line large-file bar. That is a genuine net-new finding on dxkit's
own code — the recall-drop fix in the previous commit just un-masked it (the
spurious ref-based drift had been demoting it to a warning). Fix the finding,
per dxkit's own dogfood rule, rather than weaken the comprehensive CI gate.

splitTools / toolRecall / resolveToolInputs are a cohesive unit (turn a
capability's provenance into a drift-stable RecallContext) that belongs beside
the registry, not inside it. Moved to producers/recall-inputs.ts; the registry
imports them. index.ts is back to 492 lines and reads as a listing of producers
again. buildToolsMap's Rule 19 allowlist gains the new file (it is still recall
plumbing, just in its own module).
