+++ src/baseline/producers/index.ts
@@ -57,7 +57,7 @@ import type { InlineAllowlistOccurrence } from '../../allowlist/gather';
 import type { CustomCheckFinding } from '../../analyzers/custom-checks/types';
 import type { BaselineEntry, RichBaselineEntry } from '../types';
 import { RECALL_EPOCHS, type RecallContext, type RecallMap } from '../recall';
-import { buildToolsMap } from '../tool-versions';
+import { resolveToolInputs, splitTools, toolRecall } from './recall-inputs';
 import { customCheckFindingsToBaselineEntries } from './custom-checks';
 import { largeFilesToBaselineEntries } from './health';
 import { duplicationToBaselineEntries, staleFilesToBaselineEntries } from './quality';
@@ -182,49 +182,6 @@ export interface BaselineProducer {
    * duplicating them here would double-report one cause.
    */
   readonly recallContexts: (ctx: ProducerContext) => ReadonlyMap<IdentityKind, RecallContext>;
-}
-
-/** Split a provenance `tool` field back into individual tool names. A
- *  capability's provenance is a `uniqueJoin(', ')` of every provider that
- *  contributed (e.g. `'gitleaks, grep-secrets'` when both ran), so each name
- *  resolves its own version rather than the joined string being recorded as
- *  one unversioned tool. */
-function splitTools(joined: string | null | undefined): string[] {
-  if (!joined) return [];
-  return joined
-    .split(',')
-    .map((s) => s.trim())
-    .filter(Boolean);
-}
-
-/** One `RecallContext` from a set of tool names, resolving each to its
- *  version. `buildToolsMap` handles in-process scanners (`grep-secrets`,
- *  `tls-bypass-registry`), tagging them with the dxkit version so a dxkit
- *  upgrade invalidates the kinds those scanners feed — preserving exactly the
- *  drift semantics the pre-Rule-19 `toolchainHash` had, now per kind. */
-function toolRecall(kind: IdentityKind, names: readonly string[], cwd: string): RecallContext {
-  return { epoch: RECALL_EPOCHS[kind], inputs: resolveToolInputs(names, cwd) };
-}
-
-/**
- * Resolve tool names to versions, dropping the ones that resolve to `unknown`.
- *
- * `unknown` means the name is not a registry tool at all — a builtin like
- * `find` or `git`, which every gather legitimately uses and which `findTool`
- * has no version for. It is a CONSTANT, so it can never discriminate one run
- * from another: as a recall input it is pure noise, and it would pollute the
- * `tools` map that `baseline show` renders (the invariant D143 pinned: no tool
- * in that map reads `unknown`).
- *
- * `present` is kept, deliberately. It means the tool IS installed but its
- * version probe came back empty, and `present -> 8.24.0` on a later run is a
- * real change in what we know about the scanner.
- */
-function resolveToolInputs(names: readonly string[], cwd: string): Record<string, string> {
-  const resolved = buildToolsMap([...new Set(names)].sort(), cwd);
-  return Object.fromEntries(
-    Object.entries(resolved).filter(([, version]) => version !== 'unknown'),
-  );
 }
 
 /**
